requirements.txt
flask>=3.1.0
flask-cors>=4.0.1
flask-sqlalchemy>=3.1.1
marshmallow>=3.21.3
sqlalchemy>=2.0.36
pytest>=8.2.2
pytest-flask>=1.3.0
python-dotenv>=1.0.1
requests>=2.31.0
PyJWT>=2.8.0
gunicorn>=21.2.0